requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
altair==4.2.2